refactor: 도서 상세페이지 코드 분리 - header, content
